move Density to correct package
